==> analysis/02_track_data.ipynb <==
# Notebook 2: Get Additional Music Data from Spotify API

### Introduction

Now that I have the list of 1000 albums to get data for, I can start working with the Spotify API to get the additional data required. I'll need to first get album IDs and artist genre which I will use to validate my PCA results later. Then I can get the track list for each album and the individual audio features for each track. 

You'll need a Spotify token of your own. I have a separate file called credentials.py in the same folder that uses the util function from the spotipy library to create a token with my username, client_id, client_secret, redirect_uri, and scope. You can find more details on that here, under "Becoming a Spotify Developer" : https://towardsdatascience.com/get-your-spotify-streaming-history-with-python-d5a208bbcbd3

import pandas as pd
import pickle
import sys
import requests
from credentials import token

sys.setrecursionlimit(1000000) #to allow pickling

### Read in Critics DataFrame

The critics dataframe contains information on the artist and album that I'll need to extract, format, and feed into the Spotify API.

import pickle
with open('../data/critics_df_all.pickle', 'rb') as read_file:
    critics = pickle.load(read_file)

critics.info()

album_title = critics.Album_Title
artist_name = critics.Artist_Name

#This is the format to submit information to the Spotify API.
q_list = list(map(lambda x,y: 'album:'+ str(x) + ' artist:' + str(y), album_title, artist_name))

# with open('../data/album_artist_list.pickle', 'wb') as to_write:
#     pickle.dump(q_list, to_write)

### Get Album IDs for Reviewed Albums

# with open('../data/album_artist_list.pickle', 'rb') as read_file:
#     q_list = pickle.load(read_file)

def get_info(q, token, type = 'album', query = 'id'):
    headers = {
    'Accept': 'application/json',
    'Content-Type': 'application/json',
    'Authorization': f'Bearer ' + token,
    }
    params = [
        ('q', q),
        ('type', type),
        ('market', 'US')
    ]
    json_key = str(type + 's')
    try: 
        response = requests.get('https://api.spotify.com/v1/search',
                           headers = headers, params = params, timeout = 5)
        json = response.json()
        first_result = json[json_key]['items'][0][query]
        return first_result
    except:
        return "None Found"

album_ids = pd.DataFrame([q_list, [get_info(i, token) for i in q_list]]).T

album_ids.columns = ['album_artist', 'album_id']

album_ids

### Get Artist Genre Data

def get_artist_id(q, token):
    headers = {
    'Accept': 'application/json',
    'Content-Type': 'application/json',
    'Authorization': f'Bearer ' + token,
    }
    params = [
        ('q', q),
        ('type', 'artist'),
        ('market', 'US')
    ]
    try: 
        response = requests.get('https://api.spotify.com/v1/search',
                           headers = headers, params = params, timeout = 5)
        json = response.json()
        first_result = json['artists']['items'][0]['genres']
        return first_result
    except:
        return "None Found"

album_genre = pd.concat([album_ids, pd.Series([get_info(i, token, type = 'artist', query = 'genres') 
                                 for i in artist_name]).rename('genre', inplace = True)], axis = 1)

album_genre.info()

# This drops all instances where the album from Metacritic is not on Spotify, as well as 
# instances when the same album shows up twice. 
album_genre = album_genre[album_genre.album_id != 'None Found'].drop_duplicates(subset = 'album_id').reset_index(drop = True)

album_genre.info()

# with open('../data/album_genre_critic.pickle', 'wb') as to_write:
#     pickle.dump(album_genre, to_write)

### Get Track List for each Album ID

# with open('../data/album_genre_critic.pickle', 'rb') as read_file:
#     album_genre = pickle.load(read_file)

def get_tracklist(album_id, token):
    headers = {
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': f'Bearer ' + token,
    }
    get_url = 'https://api.spotify.com/v1/albums/' + album_id + '/tracks'
    
    response = requests.get(get_url, headers=headers, params=None)
    json = response.json()
    first_result = []
    album_id_list = []
    for i in json['items']:
        first_result.append(i['id'])
        album_id_list.append(album_id)
    return pd.DataFrame(list(zip(album_id_list, first_result)), columns = ['album_id', 'track_id'])

validalbum_trackid = pd.concat([get_tracklist(i, token)  for i in album_genre['album_id']])

album_track_df = album_genre.merge(validalbum_trackid, how = 'inner', on = 'album_id')

# with open('../data/album_track_df.pickle', 'wb') as to_write:
#     pickle.dump(album_track_df, to_write)

### Get Track Features for All Tracks

# with open('../data/album_track_df.pickle', 'rb') as read_file:
#     album_track_df = pickle.load(read_file)

tracks = album_track_df['track_id']

def get_features(track_list, token):
    headers = {
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': f'Bearer ' + token,
    }
    ids = ','.join(track_list)    
    params = [
        ('ids', ids),
        ('market', 'US')
    ]    
    response = requests.get('https://api.spotify.com/v1/audio-features/', headers=headers, params=params)
    json = response.json()
    return pd.DataFrame(json['audio_features'])

all = pd.concat([pd.concat(list(map(lambda x, y: pd.concat([album_track_df.iloc[x:y, :].reset_index(drop = True), 
                                                   get_features(tracks[x:y], token)], axis = 1), 
                           [i for i in range(0, 12101, 50)], 
                           [i for i in range(50, 12151, 50)]))), 

                 pd.concat([album_track_df.iloc[12150:, :].reset_index(drop = True), 
                            get_features(tracks[12150:], token)], axis = 1)]).reset_index(drop = True)

all.info()

# with open('../data/full.pickle', 'wb') as to_write:
#     pickle.dump(all, to_write)

### Explore all features

# with open('../data/full.pickle', 'rb') as read_file:
#     all = pickle.load(read_file)

all.album_id.value_counts().value_counts()
# Albums have, at most, 20 tracks. 

all.describe()

The ranges of values for danceability, energy, key, loudness, mode, speechiness, acousticness, instrumentalness, liveness, valence, duration_ms all make sense.  
There is no track that is extremely danceable, loud, 100% speechy, confidently acoustic, confidently instrumental, or with totally positive valence.  
67% of all tracks are in major key.  
Tracks are generally lower valence.  
There are a few songs that are mostly spoken words, but the majority are not speechy.  
Most tracks are not live.  
Tempo ranges from 0 bpm (???) to 220 (makes sense).  
Mean track duration is around 4 minutes, with a minimum of 4 seconds, and a maximum of 7 minutes. Median is almost 4 minutes.  
Not sure what it means to have a 0 time signature, but most tracks are in 4/4 time.
