==> analysis/04_genre_reduction.ipynb <==
# Notebook 4: Dimensionality Reduction of Genre

### Introduction

I need to clean up the genre information for each artist, since I plan to validate the results of the PCA by looking at how the albums cluster, expecting albums with similar genres to cluster near each other. Currently, there are multiple genres assigned for each artist, so I will do dimensionality reduction using CorEx to assign each album to a genre.

import pandas as pd
import numpy as np
import pickle
import sys

sys.setrecursionlimit(1000000) #to allow pickling

from corextopic import corextopic as ct
from sklearn.feature_extraction.text import CountVectorizer

with open('../data/full.pickle', 'rb') as read_file:
    all = pickle.load(read_file)

all.info()

### Process Genre Data To Prepare for Count Vectorization

genre_data = all.loc[:, ['album_artist', 'genre']]
genre_data['genre_flat'] = genre_data.genre.map(lambda x: ' '.join([i.replace(' ', '_').replace('-', '_') for i in x]))
genre_data['genre_flat2'] = genre_data.genre.map(lambda x: ' '.join([i.replace(' ', '_').replace('-', '_').replace('+', '').replace("'", '').replace('&', '') for i in x]))

genre_album = genre_data.loc[:, ['album_artist', 'genre_flat2']].drop_duplicates().reset_index(drop = True)

### Use Count Vectorizer to Get All Possible Genres

all_genres = CountVectorizer()
album_genres = all_genres.fit_transform(genre_album['genre_flat2'])

Note that applying this count vectorizer to different kinds of music may not yield good results, because the available genres will change dramatically with the inclusion of new kinds of music.

#List of genres detected
all_genres.get_feature_names()[:10]

#Dataframe of all albums and which genres they are associated with
#pd.DataFrame(album_genres.toarray(), columns = all_genres.get_feature_names())

### Use CorEx for Dimensionality Reduction

genre_category = ct.Corex(n_hidden = 10, words = all_genres.get_feature_names(), seed = 1)
genre_category.fit(album_genres, words = all_genres.get_feature_names(), docs = genre_album['genre_flat2'])

genre_name = genre_category.get_topics()
for n,genre_name in enumerate(genre_name):
    genre_words,_ = zip(*genre_name)
    print('{}: '.format(n) + ','.join(genre_words))

### Assign Genres to Each Album By Probability, With an "Other" Category

album_assigned_genre = pd.concat([genre_album.iloc[:, 0], 
           pd.DataFrame(genre_category.predict(album_genres), 
                        columns = ['genre'+str(i) for i in range(10)])],  axis = 1)

probabilities = pd.DataFrame(genre_category.p_y_given_x, columns = ['genre'+str(i) for i in range(10)])

genres_assigned = album_assigned_genre.set_index('album_artist')
genres_assigned['any_genre'] = genres_assigned.apply(lambda x: 1 if x.sum() >= 1 else 0, axis = 1)
genres_assigned['genre10'] = genres_assigned.any_genre.map(lambda x: True if x == 0 else False) #genre9 is catchall
genres_assigned = genres_assigned.reset_index()
#In instances where the album fits in multiple genres, assign the one with the highest probability
genres_assigned['genre_num'] = probabilities.apply(lambda x: np.argmax(x), axis = 1)
#If album does not fit into any genre (probability < 50%), put it in "Other" category
genres_assigned['final_genre'] = genres_assigned.apply(lambda x: 10 if x.genre10 else x.genre_num, axis = 1)

### Name Genres, Finalize Genre Data

album_assigned = genres_assigned[['album_artist', 'final_genre']]

genre_mapping = pd.DataFrame({0: 'Folk-SingerSongwriter', 
                              1: 'HipHop-Rap',
                              2: 'Dance-Electronica',
                              3: 'Metal-ClassicRock', 
                              4: 'Indie',
                              5: 'Experimental',
                              6: 'AlternativeRnB-IndieSoul',
                              7: 'Rock-NewWave',
                              8: 'Electropop-Pop',
                              9: 'ArtRock-SlowCore',
                              10: 'Other'}, index = [1]).T.reset_index()

album_genre_final = album_assigned.merge(genre_mapping, left_on = 'final_genre', right_on = 'index')[['album_artist', 1]]
album_genre_final.columns = ['album_artist', 'genre']

album_genre_final.genre.value_counts()

# with open('../data/album_genre.pickle', 'wb') as to_write:
#     pickle.dump(album_genre_final, to_write)
