==> analysis/06_recommender.ipynb <==
# Notebook 6: Calculating Distance to Get Album Recommendations

### Introduction

At this point, I am fairly confident that I can group similar albums next to each other, and separate dissimilar albums. Now I can calculate the distances between albums to get the next most-similar album.

import pandas as pd
import numpy as np
import pickle
import sys

sys.setrecursionlimit(1000000) #to allow pickling

from sklearn.metrics.pairwise import cosine_similarity

with open('../data/components_by_album.pickle', 'rb') as read_file:
    components_by_album = pickle.load(read_file)

### Calculate Distances Between Albums Using Cosine Similarity

album_distances = (pd.concat([components_by_album.reset_index().iloc[:, 0], 
                             pd.DataFrame(cosine_similarity(components_by_album, 
                                                            components_by_album))], axis = 1)
                   .set_index('album_artist'))

album_distances

album_distances['combined'] = album_distances.values.tolist()

#max_cosine variable has the index of the closest album, that isn't itself
album_distances['max_cosine'] = album_distances.combined.map(lambda x: np.argsort(x)[::-1][1])

### Check Recommendations

I'll take a quick look at a few albums and their closest similar albums, just to make sure things make sense.

album_distances.iloc[:, 993]

album_distances.iloc[[159, 792], 993]

In broad strokes, there are similarities here. Sigur Rós's Ágætis Byrjun is similar to Centralia's Mountains - they are both very atmospheric, instrumental albums. Both Blueprint's 1988 and Jay-Z's Black Album are Hip-Hop/Rap. This makes sense.

### Return Most Similar Album

These are the five most similar albums to Blueprint's 1988.

dist = cosine_similarity(np.array(components_by_album.loc['album:1988 artist:Blueprint', ]).reshape(1, -1), components_by_album)[0]

match = np.argsort(dist)[::-1][:10]

components_by_album.reset_index().iloc[match[1:6], 0].values

The five most similar albums to The Breeders' All Nerve include Lykke Li's Wounded Rhymes.

match = (np.argsort(cosine_similarity(np
                                      .array(components_by_album
                                             .loc['album:All Nerve artist:The Breeders', ])
                                      .reshape(1, -1), components_by_album)[0])[::-1][:10])

components_by_album.reset_index().iloc[match[1:6], 0].values

The five most similar albums to Janelle Monáe's Dirty Computer includes Solange's A Seat at the Table.

match = (np.argsort(cosine_similarity(np
                                      .array(components_by_album
                                             .loc['album:Dirty Computer artist:Janelle Monáe', ])
                                      .reshape(1, -1), components_by_album)[0])[::-1][:10])

components_by_album.reset_index().iloc[match[1:6], 0].values

### Next Steps

This recommendation system will be implemented in a Flask app that the user can interact with to get album recommendations based on an input album. Please see the main repo page for more details.
